# Box precomputes and uses the DH shared key.