--- a/src/main/java/org/abstractj/kalium/crypto/Box.java
+++ b/src/main/java/org/abstractj/kalium/crypto/Box.java
@@ -16,6 +16,7 @@
 
 package org.abstractj.kalium.crypto;
 
+import org.abstractj.kalium.NaCl;
 import org.abstractj.kalium.encoders.Encoder;
 import org.abstractj.kalium.keys.PrivateKey;
 import org.abstractj.kalium.keys.PublicKey;
@@ -36,14 +37,15 @@ import static org.abstractj.kalium.crypto.Util.removeZeros;
  */
 public class Box {
 
-    private final byte[] privateKey;
-    private final byte[] publicKey;
+    private final byte[] sharedKey;
 
     public Box(byte[] publicKey, byte[] privateKey) {
-        this.publicKey = publicKey;
-        this.privateKey = privateKey;
         checkLength(publicKey, PUBLICKEY_BYTES);
         checkLength(privateKey, SECRETKEY_BYTES);
+
+        sharedKey = new byte[NaCl.Sodium.BEFORENMBYTES];
+        isValid(sodium().crypto_box_curve25519xsalsa20poly1305_beforenm(
+                sharedKey, publicKey, privateKey), "Key agreement failed");
     }
 
     public Box(PublicKey publicKey, PrivateKey privateKey) {
@@ -58,8 +60,8 @@ public class Box {
         checkLength(nonce, NONCE_BYTES);
         byte[] msg = prependZeros(ZERO_BYTES, message);
         byte[] ct = new byte[msg.length];
-        isValid(sodium().crypto_box_curve25519xsalsa20poly1305(ct, msg,
-                msg.length, nonce, publicKey, privateKey), "Encryption failed");
+        isValid(sodium().crypto_box_curve25519xsalsa20poly1305_afternm(ct, msg,
+                msg.length, nonce, sharedKey), "Encryption failed");
         return removeZeros(BOXZERO_BYTES, ct);
     }
 
@@ -71,8 +73,9 @@ public class Box {
         checkLength(nonce, NONCE_BYTES);
         byte[] ct = prependZeros(BOXZERO_BYTES, ciphertext);
         byte[] message = new byte[ct.length];
-        isValid(sodium().crypto_box_curve25519xsalsa20poly1305_open(message, ct,
-                message.length, nonce, publicKey, privateKey), "Decryption failed. Ciphertext failed verification.");
+        isValid(sodium().crypto_box_curve25519xsalsa20poly1305_open_afternm(
+                        message, ct, message.length, nonce, sharedKey),
+                "Decryption failed. Ciphertext failed verification.");
         return removeZeros(ZERO_BYTES, message);
     }